refactor(switch,minimald,minvmd): unify gvproxy switch primitives in a crate
The gvproxy switch subnet, MAC derivation, vsock wire constants, and
`-config` rendering had drifted into parallel copies in minimald and
minvmd (the SwitchSubnet types had even diverged: /8..=/29 vs /1..=/30
validation, Option vs non-Option accessors, separate error types).

Collapse them into one definition in a new dedicated `switch` crate —
these are network primitives, not part of the `minimald-rpc` wire
contract, so they get their own crate rather than living in the RPC crate.
Unified on minimald's design (the richer one the netns path uses):
SwitchSubnet (/8..=/29, non-Option accessors, daemon_ip), MacAddr,
GATEWAY_MAC, DEFAULT_MTU, DEFAULT_SUBNET, the VSOCK constants, and
render_gvproxy_config (tuple leases). Both crates re-export them, so
`minimald::net::*` / `minvmd::net::*` keep working.

minvmd adapts: its render drops the now-unnecessary Option handling and
its SwitchSubnet API tests (asserting the old /1,/30,Option behaviour)
are dropped — the shared type is tested in the `switch` crate.

No behaviour change for the deployed /16 subnet. Refs #581 review
(twitchyliquid64).
