+++ crates/minimald/src/net/mod.rs
@@ -31,7 +31,6 @@ pub(crate) mod gvproxy_network;
 #[cfg(feature = "networking-wg")]
 pub mod wg;
 
-use std::fmt;
 use std::io;
 use std::net::Ipv4Addr;
 use std::path::{Path, PathBuf};
@@ -41,33 +40,15 @@ use std::time::Duration;
 use tokio::process::{Child, Command};
 use tokio::sync::watch;
 
-/// Default gvproxy switch subnet: the RFC 6598 (CGNAT) `100.64.0.0/16` block.
-///
-/// Chosen so PTask switch addresses never collide with the common RFC 1918
-/// ranges a host or its containers already use. Configurable via
-/// [`SwitchSubnet`] so a deployment that already uses `100.64/10` can override
-/// it.
-pub const DEFAULT_SUBNET: SwitchSubnet = SwitchSubnet {
-    base: Ipv4Addr::new(100, 64, 0, 0),
-    prefix: 16,
+// The gvproxy-switch primitives (subnet arithmetic, MAC derivation, wire
+// constants, `-config` rendering) live in the shared `switch` crate.
+// Re-exported here so `minimald::net::{SwitchSubnet, MacAddr, …}` keeps working.
+// `::switch` (leading `::`) is the extern crate, disambiguated from the sibling
+// `crate::net::switch` relay module.
+pub use ::switch::{
+    DEFAULT_MTU, DEFAULT_SUBNET, GATEWAY_MAC, InvalidPrefix, MacAddr, SwitchSubnet,
+    VSOCK_GVPROXY_SHUTTLE_PORT, VSOCK_HOST_CID, render_gvproxy_config,
 };
-
-/// MTU advertised to the switch and the tap devices. gvproxy's own default.
-pub const DEFAULT_MTU: u16 = 1500;
-
-/// Stable, locally-administered MAC for the switch gateway.
-pub const GATEWAY_MAC: MacAddr = MacAddr([0x5a, 0x94, 0xef, 0xe4, 0x0c, 0xdd]);
-
-/// AF_VSOCK CID of the host as seen from inside a libkrun guest. Well-known:
-/// `VMADDR_CID_HOST == 2`. The per-PTask shuttle dials this CID to reach the
-/// host gvproxy switch (DM1/3/4).
-pub const VSOCK_HOST_CID: u32 = 2;
-
-/// vsock port the per-PTask shuttle connects to on the host to reach the host
-/// gvproxy switch. Must match `minvmd`'s
-/// `net::VSOCK_GVPROXY_SHUTTLE_PORT`; libkrun bridges this guest-initiated
-/// connection to the host gvproxy `-listen` socket.
-pub const VSOCK_GVPROXY_SHUTTLE_PORT: u32 = 1024;
 
 /// How the per-host gvproxy switch is reached, selected by deployment model.
 ///
@@ -103,14 +84,6 @@ pub enum NetError {
     /// The configured subnet has no remaining host address to hand out.
     #[error("gvproxy subnet {0} is exhausted; no free PTask address remains")]
     SubnetExhausted(SwitchSubnet),
-    /// A subnet was constructed with a prefix outside the supported `8..=29`
-    /// range — a configuration error, distinct from runtime address exhaustion.
-    #[error(
-        "gvproxy subnet prefix /{0} is invalid; must be in 8..=29 (narrower has \
-         no room for a PTask address; wider lets the high octet vary, which the \
-         derived MAC cannot keep collision-free)"
-    )]
-    InvalidPrefix(u8),
     /// Spawning the gvproxy binary failed.
     #[error("spawning gvproxy at {path:?}: {source}")]
     Spawn {
@@ -134,140 +107,6 @@ pub enum NetError {
     /// An I/O error not attributable to a more specific failure mode.
     #[error(transparent)]
     Io(#[from] io::Error),
-}
-
-/// A 48-bit Ethernet MAC address.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct MacAddr(pub [u8; 6]);
-
-impl fmt::Display for MacAddr {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let [a, b, c, d, e, g] = self.0;
-        write!(f, "{a:02x}:{b:02x}:{c:02x}:{d:02x}:{e:02x}:{g:02x}")
-    }
-}
-
-impl MacAddr {
-    /// Derives a stable, locally-administered unicast MAC for a switch IP.
-    ///
-    /// Uses the QEMU OUI `52:54:00` (locally administered) followed by the
-    /// low three octets of the address. Within a single `/16` (or narrower)
-    /// switch subnet the low three octets are unique, so the derived MAC is
-    /// collision-free and deterministic — minimald can pre-seed gvproxy's
-    /// static-lease table without round-tripping through DHCP.
-    #[must_use]
-    pub fn for_switch_ip(ip: Ipv4Addr) -> Self {
-        let o = ip.octets();
-        Self([0x52, 0x54, 0x00, o[1], o[2], o[3]])
-    }
-}
-
-/// An IPv4 subnet (`base`/`prefix`) the switch hands addresses out of.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct SwitchSubnet {
-    base: Ipv4Addr,
-    prefix: u8,
-}
-
-impl fmt::Display for SwitchSubnet {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}/{}", self.network(), self.prefix)
-    }
-}
-
-impl Default for SwitchSubnet {
-    fn default() -> Self {
-        DEFAULT_SUBNET
-    }
-}
-
-impl SwitchSubnet {
-    /// Builds a subnet from a base address and prefix length.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`NetError::InvalidPrefix`] for a prefix outside `8..=29`. A
-    /// prefix narrower than /29 has no room for the reserved
-    /// network/gateway/host-alias/broadcast addresses plus a PTask address. A
-    /// prefix wider than /8 lets the high octet vary, which
-    /// [`MacAddr::for_switch_ip`] does not fold into the derived MAC, so two
-    /// addresses differing only in that octet would collide.
-    pub fn new(base: Ipv4Addr, prefix: u8) -> Result<Self, NetError> {
-        // Reserve four addresses (network, gateway, host-alias, broadcast): a
-        // /29 (8 addresses) leaves four allocatable hosts, a /30 (4) leaves
-        // none, so anything narrower than /29 has no room for a PTask. The lower
-        // bound keeps MacAddr::for_switch_ip collision-free — it derives the MAC
-        // from the low three octets, so the high octet must be pinned by the
-        // prefix (/8 or narrower). An out-of-range prefix is a misconfiguration,
-        // not runtime exhaustion, so it is reported as InvalidPrefix.
-        if !(8..=29).contains(&prefix) {
-            return Err(NetError::InvalidPrefix(prefix));
-        }
-        Ok(Self { base, prefix })
-    }
-
-    fn mask(self) -> u32 {
-        if self.prefix == 0 {
-            0
-        } else {
-            u32::MAX << (32 - self.prefix)
-        }
-    }
-
-    /// The prefix length (the `/N` of the subnet), used to render a PTask's
-    /// address as CIDR when configuring its tap.
-    #[must_use]
-    pub fn prefix(self) -> u8 {
-        self.prefix
-    }
-
-    /// The network address (host bits zeroed).
-    #[must_use]
-    pub fn network(self) -> Ipv4Addr {
-        Ipv4Addr::from(u32::from(self.base) & self.mask())
-    }
-
-    /// The broadcast address (host bits set).
-    #[must_use]
-    pub fn broadcast(self) -> Ipv4Addr {
-        Ipv4Addr::from(u32::from(self.network()) | !self.mask())
-    }
-
-    /// The gateway address: the first usable host (`network + 1`).
-    #[must_use]
-    pub fn gateway(self) -> Ipv4Addr {
-        Ipv4Addr::from(u32::from(self.network()) + 1)
-    }
-
-    /// The host alias address (`broadcast - 1`): a virtual IP that gvproxy NATs
-    /// to the host loopback so a PTask can reach host services. Reserved, never
-    /// handed to a PTask.
-    #[must_use]
-    pub fn host_alias(self) -> Ipv4Addr {
-        Ipv4Addr::from(u32::from(self.broadcast()) - 1)
-    }
-
-    /// The daemon address (`broadcast - 2`): the guest root netns' primary tap,
-    /// which gives `minimald` itself egress through the host gvproxy (so it can
-    /// fetch upstream packages). Reserved from the top so the PTask range still
-    /// starts at `network + 2` and never collides with it.
-    #[must_use]
-    pub fn daemon_ip(self) -> Ipv4Addr {
-        Ipv4Addr::from(u32::from(self.broadcast()) - 2)
-    }
-
-    /// The first address that may be allocated to a PTask (`network + 2`,
-    /// i.e. the first host after the gateway).
-    fn first_ptask(self) -> u32 {
-        u32::from(self.network()) + 2
-    }
-
-    /// The last address that may be allocated to a PTask (`broadcast - 3`),
-    /// leaving the daemon address at `broadcast - 2` and the host alias at
-    /// `broadcast - 1` reserved.
-    fn last_ptask(self) -> u32 {
-        u32::from(self.broadcast()) - 3
-    }
 }
 
 /// One PTask's place on the switch: a never-reused IP and its derived MAC.
@@ -347,45 +186,6 @@ impl IpAllocator {
     }
 }
 
-/// Renders the gvproxy YAML config for the given subnet and static leases.
-///
-/// gvproxy v0.8.9 has no `-subnet` CLI flag: the subnet, gateway, NAT alias,
-/// and DHCP static leases are all expressed through a `-config` YAML file (see
-/// the spike). minimald owns the allocation table and writes it here before
-/// every (re)start so the switch's static leases match minimald's address
-/// book.
-///
-/// gvproxy reads this file only at spawn time, so a rewrite triggered by a
-/// later [`attach`](SwitchClient::attach) takes effect only on the next switch
-/// (re)start. That is intentional: an `OwnIp` PTask configures its switch
-/// address statically (the spike's static-lease recipe) rather than via DHCP,
-/// so `dhcpStaticLeases` is a startup-time seed, not a live source a running
-/// gvproxy must re-read for each new PTask.
-#[must_use]
-pub fn render_gvproxy_config(subnet: SwitchSubnet, leases: &[PtaskLease]) -> String {
-    let mut s = String::with_capacity(256 + leases.len() * 48);
-    s.push_str("stack:\n");
-    s.push_str(&format!("  mtu: {DEFAULT_MTU}\n"));
-    s.push_str(&format!("  subnet: \"{subnet}\"\n"));
-    s.push_str(&format!("  gatewayIP: \"{}\"\n", subnet.gateway()));
-    s.push_str(&format!("  gatewayMacAddress: \"{GATEWAY_MAC}\"\n"));
-    s.push_str("  nat:\n");
-    s.push_str(&format!("    \"{}\": \"127.0.0.1\"\n", subnet.host_alias()));
-    s.push_str("  gatewayVirtualIPs:\n");
-    s.push_str(&format!("    - \"{}\"\n", subnet.host_alias()));
-    s.push_str("  dhcpStaticLeases:\n");
-    if leases.is_empty() {
-        // gvproxy accepts an empty map; keep the key present for clarity.
-        s.push_str("    {}\n");
-    } else {
-        for lease in leases {
-            s.push_str(&format!("    \"{}\": \"{}\"\n", lease.ip, lease.mac));
-        }
-    }
-    s
-}
-
-/// Supervises the single per-host gvproxy switch process (R1.4).
 ///
 /// The switch is reference-counted against the set of attached `OwnIp` PTasks:
 /// it is spawned lazily on the first attach and torn down after the last
@@ -529,7 +329,13 @@ impl SwitchClient {
                 source,
             })?;
         let path = self.config_path();
-        let body = render_gvproxy_config(self.allocator.subnet(), self.allocator.leases());
+        let leases: Vec<_> = self
+            .allocator
+            .leases()
+            .iter()
+            .map(|l| (l.ip, l.mac))
+            .collect();
+        let body = render_gvproxy_config(self.allocator.subnet(), &leases);
         tokio::fs::write(&path, body)
             .await
             .map_err(|source| NetError::WriteConfig { path, source })
@@ -757,7 +563,7 @@ mod tests {
     fn subnet_rejects_overly_narrow_prefix() {
         assert!(matches!(
             SwitchSubnet::new(Ipv4Addr::new(10, 0, 0, 0), 30),
-            Err(NetError::InvalidPrefix(30))
+            Err(InvalidPrefix(30))
         ));
     }
 
@@ -767,7 +573,7 @@ mod tests {
         // does not cover, so the constructor rejects it to keep MACs unique.
         assert!(matches!(
             SwitchSubnet::new(Ipv4Addr::new(10, 0, 0, 0), 7),
-            Err(NetError::InvalidPrefix(7))
+            Err(InvalidPrefix(7))
         ));
     }
 
@@ -775,7 +581,8 @@ mod tests {
     fn config_contains_subnet_gateway_and_leases() {
         let mut a = IpAllocator::new(SwitchSubnet::default());
         let lease = a.allocate().unwrap();
-        let cfg = render_gvproxy_config(a.subnet(), a.leases());
+        let leases: Vec<_> = a.leases().iter().map(|l| (l.ip, l.mac)).collect();
+        let cfg = render_gvproxy_config(a.subnet(), &leases);
         assert!(cfg.contains("subnet: \"100.64.0.0/16\""));
         assert!(cfg.contains("gatewayIP: \"100.64.0.1\""));
         assert!(cfg.contains(&format!("\"{}\": \"{}\"", lease.ip, lease.mac)));
