+++ crates/switch/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "switch"
+version = "0.0.1"
+edition.workspace = true
+publish.workspace = true
+
+[dependencies]
